Improve transcription API and handle missing configuration
